Fix telegram crash when requesting report for invalid addr
If attempting to request a report for a miner that doesn't exist on the
pool the telegram bot crashes (`data.stats` is undefined, leading to an
exception when calling `data.stats.hashrate`).

This adds a new "not found" error message if the lookup doesn't return
any stats.

Additionally this fixes an error in the event of a harder lookup failure
(e.g. from an error connecting to the api) not using the correct
variable for the recipient.